delete downstream when deletes the clone source

+++ pkg/background/generate/cleanup.go
@@ -2,38 +2,78 @@ package generate
 
 import (
 	"context"
+	"fmt"
 
 	kyvernov1 "github.com/kyverno/kyverno/api/kyverno/v1"
 	kyvernov1beta1 "github.com/kyverno/kyverno/api/kyverno/v1beta1"
-	"github.com/kyverno/kyverno/pkg/config"
+	"github.com/kyverno/kyverno/pkg/background/common"
 	"go.uber.org/multierr"
 	apierrors "k8s.io/apimachinery/pkg/api/errors"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 )
 
-func (c *GenerateController) handleGeneratePolicyRuleAbsence(ur *kyvernov1beta1.UpdateRequest) (err error) {
+func (c *GenerateController) deleteDownstream(policy kyvernov1.PolicyInterface, ur *kyvernov1beta1.UpdateRequest) (err error) {
 	if !ur.Spec.DeleteDownstream {
 		return nil
 	}
 
-	c.log.V(4).Info("policy no longer exists, deleting the update request and respective resource based on synchronize", "ur", ur.Name, "policy", ur.Spec.Policy)
-	var errs []error
-	failedDownstreams := []kyvernov1.ResourceSpec{}
-	for _, e := range ur.Status.GeneratedResources {
-		if err := c.client.DeleteResource(context.TODO(), e.GetAPIVersion(), e.GetKind(), e.GetNamespace(), e.GetName(), false); err != nil && !apierrors.IsNotFound(err) {
-			failedDownstreams = append(failedDownstreams, e)
-			errs = append(errs, err)
+	// handle data policy/rule deletion
+	if ur.Status.GeneratedResources != nil {
+		c.log.V(4).Info("policy/rule no longer exists, deleting the downstream resource based on synchronize", "ur", ur.Name, "policy", ur.Spec.Policy, "rule", ur.Spec.Rule)
+		var errs []error
+		failedDownstreams := []kyvernov1.ResourceSpec{}
+		for _, e := range ur.Status.GeneratedResources {
+			if err := c.client.DeleteResource(context.TODO(), e.GetAPIVersion(), e.GetKind(), e.GetNamespace(), e.GetName(), false); err != nil && !apierrors.IsNotFound(err) {
+				failedDownstreams = append(failedDownstreams, e)
+				errs = append(errs, err)
+			}
 		}
+
+		if len(errs) != 0 {
+			c.log.Error(multierr.Combine(errs...), "failed to clean up downstream resources on policy deletion")
+			_, err = c.statusControl.Failed(ur.GetName(),
+				fmt.Sprintf("failed to clean up downstream resources on policy deletion: %v", multierr.Combine(errs...)),
+				failedDownstreams)
+		} else {
+			c.statusControl.Success(ur.GetName(), nil)
+		}
+		return
 	}
 
-	if len(errs) != 0 {
-		c.log.Error(multierr.Combine(errs...), "failed to clean up downstream resources on policy deletion, retrying")
-		ur.Status.GeneratedResources = failedDownstreams
-		ur.Status.State = kyvernov1beta1.Failed
-		_, err = c.kyvernoClient.KyvernoV1beta1().UpdateRequests(config.KyvernoNamespace()).UpdateStatus(context.TODO(), ur, metav1.UpdateOptions{})
-	} else {
-		ur.Status.State = kyvernov1beta1.Completed
-		_, err = c.kyvernoClient.KyvernoV1beta1().UpdateRequests(config.KyvernoNamespace()).UpdateStatus(context.TODO(), ur, metav1.UpdateOptions{})
+	// handle clone source deletion
+	return c.deleteDownstreamForClone(policy, ur)
+}
+
+func (c *GenerateController) deleteDownstreamForClone(policy kyvernov1.PolicyInterface, ur *kyvernov1beta1.UpdateRequest) (err error) {
+	if !ur.Spec.DeleteDownstream {
+		return nil
+	}
+
+	for _, rule := range policy.GetSpec().Rules {
+		if ur.Spec.Rule != rule.Name {
+			continue
+		}
+
+		downstreams, err := FindDownstream(c.client, policy, rule)
+		if err != nil {
+			return err
+		}
+
+		var errs []error
+		failedDownstreams := []kyvernov1.ResourceSpec{}
+		for _, downstream := range downstreams.Items {
+			if err := c.client.DeleteResource(context.TODO(), downstream.GetAPIVersion(), downstream.GetKind(), downstream.GetNamespace(), downstream.GetName(), false); err != nil && !apierrors.IsNotFound(err) {
+				failedDownstreams = append(failedDownstreams, common.ResourceSpecFromUnstructured(&downstream))
+				errs = append(errs, err)
+			}
+		}
+		if len(errs) != 0 {
+			c.log.Error(multierr.Combine(errs...), "failed to clean up downstream resources on source deletion")
+			_, err = c.statusControl.Failed(ur.GetName(),
+				fmt.Sprintf("failed to clean up downstream resources on source deletion: %v", multierr.Combine(errs...)),
+				failedDownstreams)
+		} else {
+			c.statusControl.Success(ur.GetName(), nil)
+		}
 	}
 	return
 }
